LUTECE-1870 : intregrate the base url in the page cache key
This covers all usages of the base_url in the pages content, and does
not impose a strict form on the base html tag.
The method to substitute the base_url in the base tag is removed.

The impact should be minimal when the site is accessed by only one URL,
and cache size usage should be bigger when accessed by multiple URLs.

A test is added to check that cached pages contain the correct base url.