LUTECE-1889 : Stop storing an XPageApplication instance in the XPageApplicationEntry
There's one instance of xpage per session, so this instance is of no use.
Moreover, when xpages will be spring managed, it will not be possible to
retreive an instance when registering the XPageApplicationEntry because
it does not happen in a request context.

StandaloneAppService is updated to not use the instance stored in the entry, which
might break apps relying on xpage *not* being per session and using the
StandaloneAppService. This can be dealt with by using a spring managed
xpageapplication with singleton scope, if necessary.