LUTECE-1946 : Securely store BO user passwords
Change the method to store user passwords to PBKDF2WithHmacSHA1.
This algorithm is chosen because it is resonably secure and available
with the standard Java JRE.
The default iteration count was chosen so that password checks take approximately
300ms on my machine.

The possiblility to select another method to store passwords is removed. Nevertheless,
old passwords are still read and are migrated to the new format on first login.
To accomplish this, the method used to store the password is prepended to the
password itself in the database.

Thus migrating away from PBKDF2WithHmacSHA1 to a more secure method in the future
can be done by using the same method, without disrupting the users.

Two new interfaces are introduced :
 * IPassword represents a stored password, and allows to check the user supplied
   password
 * IPasswordFactory, which reads or constructs IPasswords

IPassword has no public implementaion, and a concrete implementation of the
factory is exposed via Spring, to limit the risk of storing a password
insecurely.

Tests are added for changed behaviors.