LUTECE-1967 : Change the default path in config.properties
Replace the default paths in config.properties by ones that
match any dev environnement : when tests are run from maven, the
working directory is the project base directory.
The paths will be updated with absolute ones when running the app, but
the config.properties file is already read at this stage.
The second run will read the customized paths.

This makes HtmlCleanerServiceTest#testClean pass because the
expected jtidy.properties is read.