LUTECE-1975: FIX MVCApplication responseXML() encoding
Before this commit, it would always send the HTTP headers
'Content-Type: application/xml; charset=UTF-8'
but encode the response in ISO-8859-1. Now it always uses
UTF-8 for both, so the xml documents should have the
"encoding" value of the xml prolog set to UTF-8 as well
(or absent, because UTF-8 is the default)

The UTF-8 encoding in the http header comes from the Portal.jsp
@page directive which specifies that the text file Portal.jsp is
encoded in UTF-8, but is used by tomcat to choose the response encoding
if it is not specified. If one wants to change to another encoding,
it must be done in Portal.jsp. It will affect all the site.

Also, revert responseJSON to not specify the encoding for symmetry and
to allow people to send ISO-8859-1 JSON (even if that's non standard)