Revert "LUTECE-1966: add hsqldb as a dependency (scope test) so that it is in the classpath during maven test"
This reverts commit 4550ae4cde4a42c8f4e4ad7da6767d97cc9efbfa.

This change was only allowing hsql to be used on the lutece-core
artifact, not in all the plugins. Also, since the dependency was
in scope test, it was not copie to the WEB-INF/lib directory in the
target so the jar had to be manually placed in the target so that ant
would run.
The new implementation in GLOBALPOM-14 (lutece-global-pom
0a22cd654d74434c76d7843bec64038fc060b2a2) uses optional dependencies
in the parent pom of projects.