LUTECE-2068: use Version.LATEST instead of LUCENE_VERSION_X_X_X.
Lucene Util.Version almost became irrelevent after LUCENE-5859. Let's
keep the field for backwards compatibility, even though it will not be
used anymore. Using Version.LATEST will be less confusing than a fixed
version which we need to change all the time.

This field was used to be able to do lucene upgrades without reindexing.
But we were always changing the value when upgrading anyway.