LUTECE-2087 : Add a token dependending on content to css and js links from LinksInclude
If the file is local, compute the SHA-256 hash of its content, and append it as a query
parameter to the URL of the included script or css.
While doing so, make absolute URL handling more correct (by handling https and protocol
relative URLs), and more uniform between CSS and JS files.