LUTECE-2085 : change daemons maximumPoolSize to 30 to ease migration
If ScheduledThreadCorePoolSize is greater than maximumPoolSize,
the webapp fails to start with the following Exception
    org.springframework.beans.factory.BeanCreationException: Error creating bean with name 'daemonExecutor'
        Bean instantiation via constructor failed;
        nested exception is org.springframework.beans.BeanInstantiationException: Failed to instantiate [java.util.concurrent.ThreadPoolExecutor]:
                Constructor threw exception; nested exception is java.lang.IllegalArgumentException

Before a5614a948f2c3ed32704c ( Thu Jan 29 15:25:04 2015 ), the default
value was 30. So using 20 means that websites that have overriden
ScheduledThreadCorePoolSize to the previous value will fail to start.
This is more frequent than expected..
Let's be nice to them by providing a default that doesn't crash.