LUTECE-2049 : revert actionButtons to lutece v5 implementation
This macro should not be used anymore, we are keeping it for backwards
compatibility. So let's revert its to code to the one from lutece v5
because it should still work on v6. Before this revert, some
functionnality was broken, for example if used with glyphicon it would
no longer display the icon (eg icon1="icon-plus icon-white") and the
offset parameter was ignored