LUTECE-2353: Remove type="text/javascript" from script tag
The HTML spec says for the type attribute of the script tag:
Omitting the attribute, setting it to the empty string, or setting it to a JavaScript MIME type essence match,
means that the script is a classic script, to be interpreted according to the JavaScript Script top-level
production. Classic scripts are affected by the async and defer attributes, but only when the src attribute
is set. Authors should omit the type attribute instead of redundantly setting it.

text/javascript is a JavaScript MIME type essence match, therefore type="text/javascript" should be omitted.