LUTECE-2354: Remove default role form on form tag
The ARIA spec says: WAI-ARIA is designed to provide semantic information about
objects when host languages lack native semantics for the object. (...) For
example, if an element with the functionality already exists, such as a checkbox
or radio button, use the native semantics of the host language. WAI-ARIA markup
is only intended to be used to enhance the native semantics (e.g., indicating
that the element is required with aria-required), or to change the semantics to
a different purpose from the standard functionality of the element.

the form tag implies the form role, and therefore should be omitted unless justified