LUTECE-2355: Only add id attribute to table if not empty
The HTML spec says: When specified on HTML elements, the id attribute value must
be unique amongst all the IDs in the element's tree and must contain at least
one character. The value must not contain any ASCII whitespace.

Thus if the id argument of the table macro is empty, the id attribute should
not be set.