LUT-25525: Fix core_admin_role_resource initialization
The primary key for this table is autogenerated since commit 54abb2a50da4
but this commit failed to update the init script, leaving hard coded identifiers.

This provoques conflict when creating a lot of RBACs and stepping into these identifiers, such as when
writing tests around this functionnality.

Switch to autogenrating these ids.