LUT-30963: Make SecurityHeaderService threadsafe
Use a volatile cache object, which is swapped for a new one when the cache needs to be refreshed.
Implement StartUpService to populate the cache on startup.

Replace the cached map of security header by a collection. The map was only used for debug logs in low
traffic methods.

Simplify the construction of the map of security headers by type and category.